offline tested, not sure if is actually checking time into the equation, it should be doing it from logs.csv so yes

+++ bin/utils.py
@@ -132,7 +132,8 @@ class Loader():
     """
     loads and write cards locally or from cloud
     """
-    def __init__(self, local=False):
+    def __init__(self, local=True):
+        self.local = local
         if not local:
             self.proxy = GCSproxy() 
             self.proxy.gcs_load('cards.pkl')
@@ -144,8 +145,9 @@ class Loader():
         clean env after ending
         """
         try:
-            os.system("rm cards.pkl")
-            os.system("rm logs.csv")
+            if not self.local:
+                os.system("rm cards.pkl")
+                os.system("rm logs.csv")
         except: pass
 
     def load_cards(self):
@@ -165,8 +167,9 @@ class Loader():
         """
         with open('cards.pkl', 'wb') as data:
             pickle.dump(self.cards, data)
-        self.proxy.gcs_write('logs.csv')
-        self.proxy.gcs_write('cards.pkl')
+        if not self.local:
+            self.proxy.gcs_write('logs.csv')
+            self.proxy.gcs_write('cards.pkl')
 
 class GCSproxy():
     """basic proxy to handle gcs APIs"""
